stdio: Basic functions for stdout and stderr
Simple implementation of some stdio functions for Linux. We need these
to make things like the assert macro work properly and to let unit tests
write to stdout as part of their expected behavior.

Also adds a stub version of errno that needs to be filled out with all
the error codes.