toolchain: Find Clang toolchain in LLVM directory
On Linux, versioned LLVM tools get symlinked into /usr/bin. For example
clang-9 is actually located at /usr/lib/llvm-9/bin/clang but it is
symlinked as /usr/bin/clang-9. When CMake uses the /usr/bin path, that's
also where it looks for other tools. This can lead to inconsistent
versions used; such as clang-9 and the system default llvm-ar (which is
likely a much older version).

This commit makes CMake use the real path of the Clang executable.