test: improve ENV directive for sandboxed tests
A check_program directive

    ENV: FOO=bar

is exposed as both FOO and libc_test_FOO environment variables.

This allows tests run locally to check for FOO and for sandboxed test
runners to know which variables to expose to the test program (the ones
matching libc_test_*).

This improves previous behavior in which the sandboxed runners only
checked for TEST_* variables and a test using the ENV directive had
to use that prefix itself.